(frontend) Refactor to a core module
- Puts data stuff into its own module
- Prep for ngrx/store
