+++ todo-frontend/src/app/app.module.ts
@@ -3,19 +3,19 @@ import { NgModule } from '@angular/core';
 import { HttpClientModule } from '@angular/common/http';
 import { BrowserAnimationsModule } from '@angular/platform-browser/animations';
 import { FormsModule } from '@angular/forms';
+import { CoreModule } from './core/core.module';
 
 import { MatFormFieldModule } from '@angular/material/form-field';
 import { MatInputModule } from '@angular/material/input';
 import { MatToolbarModule } from '@angular/material/toolbar';
 import { MatListModule } from '@angular/material/list';
 import { MatButtonModule } from '@angular/material/button';
-import { MatCardModule} from '@angular/material/card';
-import { MatDialogModule} from '@angular/material/dialog'
+import { MatCardModule } from '@angular/material/card';
+import { MatDialogModule } from '@angular/material/dialog';
 
 import { AppRoutingModule } from './app-routing.module';
 import { AppComponent } from './app.component';
 import { DashboardComponent } from './dashboard/dashboard.component';
-import { TodoService } from './todo.service';
 import { TodoComponent } from './todo/todo.component';
 import { TodoFormDialogComponent } from './todo-form-dialog/todo-form-dialog.component';
 import { NavComponent } from './nav/nav.component';
@@ -34,15 +34,15 @@ import { NavComponent } from './nav/nav.component';
     HttpClientModule,
     BrowserAnimationsModule,
     FormsModule,
+    CoreModule,
     MatFormFieldModule,
     MatInputModule,
     MatToolbarModule,
     MatListModule,
     MatButtonModule,
     MatCardModule,
-    MatDialogModule
+    MatDialogModule,
   ],
-  providers: [TodoService],
   bootstrap: [AppComponent],
 })
 export class AppModule {}
