+++ todo-frontend/src/app/core/core.module.ts
@@ -1,0 +1,11 @@
+import { NgModule } from '@angular/core';
+import { CommonModule } from '@angular/common';
+
+import { TodoService } from './todo/todo.service';
+
+// prettier-ignore
+@NgModule({
+  imports: [CommonModule],
+  providers: [ TodoService ]
+})
+export class CoreModule {}
